kubernetes: virtual-machines: rename non jsx files to .es6
