+++ pkg/kubernetes/scripts/virtual-machines/action-types.es6
