+++ pkg/kubernetes/scripts/virtual-machines/components/vm/createVmDialog.jsx
@@ -21,7 +21,7 @@ import cockpit from 'cockpit';
 import React from 'react';
 
 import './createVmDialog.less';
-import { preventDefault } from '../../utils.jsx';
+import { preventDefault } from '../../utils.es6';
 
 const _ = cockpit.gettext;
