+++ pkg/kubernetes/scripts/virtual-machines/components/vmi/VmiOverviewTab.jsx
@@ -24,11 +24,11 @@ import { connect } from 'react-redux';
 
 import { commonTitles } from '../../../../../machines/components/vmOverviewTab.jsx';
 
-import type { Vmi, Message as MessageType, Pod } from '../../types.jsx';
-import { kindIdPrefx } from '../../utils.jsx';
-import { getNodeName, getMemory, getCPUs, getPhase, getAge } from '../../selectors.jsx';
+import type { Vmi, Message as MessageType, Pod } from '../../types.es6';
+import { kindIdPrefx } from '../../utils.es6';
+import { getNodeName, getMemory, getCPUs, getPhase, getAge } from '../../selectors.es6';
 import { getLabels } from '../util/utils.jsx';
-import {removeVmiMessage} from '../../action-creators.jsx';
+import {removeVmiMessage} from '../../action-creators.es6';
 import OverviewTab from '../common/OverviewTab.jsx';
 import Message from '../common/Message.jsx';
 import EntityLink from '../common/EntityLink.jsx';
