+++ pkg/kubernetes/scripts/virtual-machines/types.es6
