+++ pkg/kubernetes/scripts/virtual-machines/utils.es6
